dlt-hypergraph-rag-phase1/schema-discovery/requirements.txt

dlt[postgres]>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
typing-extensions>=4.0.0
